+++ src/module/data/warfareData.mjs
@@ -25,104 +25,118 @@ export default class WarfareData extends foundry.abstract.TypeDataModel {
   static defineSchema() {
     const fields = foundry.data.fields;
     const data = {
-      commander: new fields.ForeignDocumentField(getDocumentClass('Actor'), {
-        textSearch: true,
+      commander: new fields.ForeignDocumentField(getDocumentClass("Actor"), {
+        textSearch: true, label: "KNW.Warfare.Commander.Label"
       }),
       ancestry: new fields.StringField({
-        textSearch: true,
+        textSearch: true, label: "KNW.Warfare.Ancestry"
       }),
       experience: new fields.StringField({
         choices: CONFIG.KNW.CHOICES.EXPERIENCE,
-        initial: 'regular',
+        initial: "regular",
         textSearch: true,
+        label: "KNW.Warfare.Experience.Label"
       }),
       gear: new fields.StringField({
         choices: CONFIG.KNW.CHOICES.GEAR,
-        initial: 'light',
+        initial: "light",
         textSearch: true,
+        label: "KNW.Warfare.Gear.Label"
       }),
       type: new fields.StringField({
         choices: CONFIG.KNW.CHOICES.TYPE,
-        initial: 'infantry',
+        initial: "infantry",
         textSearch: true,
+        label: "KNW.Warfare.Type.Label"
       }),
       atk: new fields.NumberField({
         required: true,
         initial: 0,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.atk.long"
       }),
       def: new fields.NumberField({
         required: true,
         initial: 10,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.def.long"
       }),
       pow: new fields.NumberField({
         required: true,
         initial: 0,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.pow.long"
       }),
       tou: new fields.NumberField({
         required: true,
         initial: 10,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.tou.long"
       }),
       mor: new fields.NumberField({
         required: true,
         initial: 0,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.mor.long"
       }),
       com: new fields.NumberField({
         required: true,
         initial: 0,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.com.long"
       }),
       attacks: new fields.NumberField({
         required: true,
         initial: 1,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.attacks.long"
       }),
       dmg: new fields.NumberField({
         required: true,
         initial: 1,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.dmg.long"
       }),
       mov: new fields.NumberField({
         required: true,
         initial: 1,
         nullable: false,
         integer: true,
+        label: "KNW.Warfare.Statistics.move.long"
       }),
       tier: new fields.NumberField({
         required: true,
         initial: 1,
         choices: CONFIG.KNW.CHOICES.TIER,
         integer: true,
+        label: "KNW.Warfare.Tier"
       }),
       size: new fields.SchemaField({
         max: new fields.NumberField({
           required: true,
           initial: 6,
-          integer: true,
+          integer: true
         }),
         value: new fields.NumberField({
           required: true,
           initial: 6,
-          integer: true,
-        }),
-      }),
+          integer: true
+        })
+      }, {label: "KNW.Warfare.Statistics.size.long"}),
       traitList: new fields.StringField({
-        initial: '',
+        initial: "",
         nullable: false,
         textSearch: true,
-      }),
+        label: "KNW.Warfare.Traits.SheetLabel"
+      })
     };
 
     return data;
@@ -138,7 +152,7 @@ export default class WarfareData extends foundry.abstract.TypeDataModel {
   get commanderName() {
     const commander = this.commander;
     if (commander) return commander.name;
-    else return game.i18n.localize('KNW.Warfare.Commander.None');
+    else return game.i18n.localize("KNW.Warfare.Commander.None");
   }
 
   /**
@@ -148,22 +162,22 @@ export default class WarfareData extends foundry.abstract.TypeDataModel {
    */
   async rollStat(stat, event) {
     const roll = game.dnd5e.dice.d20Roll({
-      parts: ['@stat'],
+      parts: ["@stat"],
       data: {
-        stat: this[stat],
+        stat: this[stat]
       },
-      title: game.i18n.format('KNW.Warfare.Statistics.Test', {
+      title: game.i18n.format("KNW.Warfare.Statistics.Test", {
         stat: game.i18n.localize(`KNW.Warfare.Statistics.${stat}.long`),
-        actorName: this.parent.name,
+        actorName: this.parent.name
       }),
-      flavor: game.i18n.format('KNW.Warfare.Statistics.Test', {
+      flavor: game.i18n.format("KNW.Warfare.Statistics.Test", {
         stat: game.i18n.localize(`KNW.Warfare.Statistics.${stat}.long`),
-        actorName: this.commander?.name ?? '',
+        actorName: this.commander?.name ?? ""
       }),
       messageData: {
-        speaker: { actor: this.parent },
+        speaker: {actor: this.parent}
       },
-      event,
+      event
     });
   }
 }