+++ src/module/sheets/organizationSheet.mjs
@@ -1,17 +1,18 @@
-import OrgDevEditor from './orgDevEditor.mjs';
+import OrgDevEditor from "./orgDevEditor.mjs";
 
 export default class OrganizationSheet extends ActorSheet {
   /** @override */
   get template() {
-    return 'modules/knw-actors/templates/organization-sheet.hbs';
+    return "modules/knw-actors/templates/organization-sheet.hbs";
   }
 
   /** @override */
   static get defaultOptions() {
     return foundry.utils.mergeObject(super.defaultOptions, {
-      classes: ['dnd5e', 'sheet', 'actor', 'organization'],
+      classes: ["dnd5e", "sheet", "actor", "organization"],
       width: 720,
       height: 680,
+      viewPermission: CONST.DOCUMENT_OWNERSHIP_LEVELS.OBSERVER
     });
   }
 
@@ -24,23 +25,23 @@ export default class OrganizationSheet extends ActorSheet {
       skills: Object.entries(this.actor.system.skills).map(([key, skill]) => {
         return {
           key,
-          label: game.i18n.localize('KNW.Organization.skills.' + key),
-          bonus: skill.bonus,
+          label: game.i18n.localize("KNW.Organization.skills." + key),
+          bonus: skill.bonus
         };
       }),
       defenses: Object.entries(this.actor.system.defenses).map(([key, def]) => {
         return {
           key,
           label: game.i18n.localize(
-            'KNW.Organization.defenses.' + key + '.Label'
+            "KNW.Organization.defenses." + key + ".Label"
           ),
           level: def.level,
           score: def.score,
-          choices: CONFIG.KNW.CHOICES[key],
+          choices: CONFIG.KNW.CHOICES[key]
         };
       }),
       powerDieIMG: this.powerDieIMG,
-      powerPool: this.getMemberPowerPool(),
+      powerPool: this.getMemberPowerPool()
     };
     return context;
   }
@@ -53,23 +54,23 @@ export default class OrganizationSheet extends ActorSheet {
     return Object.entries(this.actor.system.powerPool).map(([id, current]) => {
       const actor = game.actors.get(id);
       let value = current;
-      let tooltip = 'KNW.Organization.Powers.Take';
-      let decrement = `<a class="subtract"><i class="fas fa-minus"></i></a>`;
+      let tooltip = "KNW.Organization.Powers.Take";
+      let decrement = "<a class=\"subtract\"><i class=\"fas fa-minus\"></i></a>";
       if (value === null) {
-        value = '<i class="fa-solid fa-dice-d20"></i>';
-        tooltip = 'KNW.Organization.Powers.Roll';
-        decrement = '';
+        value = "<i class=\"fa-solid fa-dice-d20\"></i>";
+        tooltip = "KNW.Organization.Powers.Roll";
+        decrement = "";
       } else if (value === 0) {
-        value = '<i class="fa-solid fa-repeat"></i>';
-        tooltip = 'KNW.Organization.Powers.Rest';
-        decrement = '';
+        value = "<i class=\"fa-solid fa-repeat\"></i>";
+        tooltip = "KNW.Organization.Powers.Rest";
+        decrement = "";
       }
       return {
         id: id,
         name: actor?.name,
         value,
         decrement,
-        tooltip: game.i18n.localize(tooltip),
+        tooltip: game.i18n.localize(tooltip)
       };
     });
   }
@@ -83,46 +84,46 @@ export default class OrganizationSheet extends ActorSheet {
 
     const dropActor = await fromUuid(data.uuid);
     if (dropActor.pack) {
-      ui.notifications.warn('KNW.Organization.Powers.Warning.NoPackActors', {
-        localize: true,
+      ui.notifications.warn("KNW.Organization.Powers.Warning.NoPackActors", {
+        localize: true
       });
       return false;
     } else if (
-      !foundry.utils.getProperty(dropActor, 'system.attributes.prof')
+      !foundry.utils.getProperty(dropActor, "system.attributes.prof")
     ) {
-      ui.notifications.warn('KNW.Organization.Powers.Warning.NoProf', {
-        localize: true,
+      ui.notifications.warn("KNW.Organization.Powers.Warning.NoProf", {
+        localize: true
       });
       return false;
-    } else if (this.actor.system.powerPool.hasOwnProperty(dropActor.id)) {
-      ui.notifications.warn('KNW.Organization.Powers.Warning.AlreadyMember', {
-        localize: true,
+    } else if (dropActor.id in this.actor.system.powerPool) {
+      ui.notifications.warn("KNW.Organization.Powers.Warning.AlreadyMember", {
+        localize: true
       });
       return false;
     }
-    this.actor.update({ [`system.powerPool.${dropActor.id}`]: null });
+    this.actor.update({[`system.powerPool.${dropActor.id}`]: null});
   }
 
   /** @override */
   activateListeners(html) {
     super.activateListeners(html);
-    html.on('click', '.body .editScore>a', this.#editScore.bind(this));
+    html.on("click", ".body .editScore>a", this.#editScore.bind(this));
 
-    html.on('click', '.skills .label.rollable', this.#rollSkill.bind(this));
-    html.on('click', '.powerDie.rollable', this.#cyclePowerDie.bind(this));
+    html.on("click", ".skills .label.rollable", this.#rollSkill.bind(this));
+    html.on("click", ".powerDie.rollable", this.#cyclePowerDie.bind(this));
     html.on(
-      'click',
-      '.powerPoolList .edit .subtract',
+      "click",
+      ".powerPoolList .edit .subtract",
       this.#decrementPowerDie.bind(this)
     );
 
-    ContextMenu.create(this, html, '.powerPoolMember', this.powerPoolItemMenu);
+    ContextMenu.create(this, html, ".powerPoolMember", this.powerPoolItemMenu);
   }
 
   async #editScore(event) {
     const statGroup = event.currentTarget.dataset.target;
     const context = {
-      statGroup,
+      statGroup
     };
     const orgDevEditor = new OrgDevEditor(this.actor, context);
     orgDevEditor.render(true);
@@ -138,76 +139,76 @@ export default class OrganizationSheet extends ActorSheet {
       (accumulator, currentValue, currentIndex, array) => {
         return (
           accumulator +
-          ' ' +
+          " " +
           CONFIG.DND5E.skills[currentValue].label +
-          (currentIndex === array.length - 1 ? '' : ',')
+          (currentIndex === array.length - 1 ? "" : ",")
         );
       },
-      game.i18n.localize('KNW.Organization.skills.Test.AssocSkills')
+      game.i18n.localize("KNW.Organization.skills.Test.AssocSkills")
     );
 
     if (validActors.length === 0)
-      ui.notifications.warn('KNW.Organization.skills.Warning.noActors', {
-        localize: true,
+      ui.notifications.warn("KNW.Organization.skills.Warning.noActors", {
+        localize: true
       });
     else if (validActors.length === 1) {
       const useProf = await Dialog.wait({
-        title: game.i18n.format('KNW.Organization.skills.Test.Title', {
-          skill: game.i18n.localize('KNW.Organization.skills.' + stat),
+        title: game.i18n.format("KNW.Organization.skills.Test.Title", {
+          skill: game.i18n.localize("KNW.Organization.skills." + stat)
         }),
         content: `<label class="flexrow"><p class="orgUseProf">${game.i18n.localize(
-          'KNW.Organization.skills.Test.UseProf'
+          "KNW.Organization.skills.Test.UseProf"
         )}</p>
         <input class="orgTestUseProf" type="checkbox" /></label>
         <p><em>${assocSkillsText}</em></p>`,
         buttons: {
           default: {
-            icon: '<i class="fa-solid fa-floppy-disk"></i>',
-            label: game.i18n.localize('KNW.Organization.skills.Test.Roll'),
+            icon: "<i class=\"fa-solid fa-floppy-disk\"></i>",
+            label: game.i18n.localize("KNW.Organization.skills.Test.Roll"),
             callback: (html) => {
-              return html.find('.orgTestUseProf')[0].checked;
-            },
-          },
-        },
+              return html.find(".orgTestUseProf")[0].checked;
+            }
+          }
+        }
       });
       this.actor.system.rollSkillTest(stat, validActors[0], useProf, event);
     } else {
       const selectOptions = validActors.map((actor) => ({
         memberID: actor.id,
-        memberName: actor.name,
+        memberName: actor.name
       }));
 
       const testInput = await Dialog.wait({
-        title: game.i18n.format('KNW.Organization.skills.Test.Title', {
-          skill: game.i18n.localize('KNW.Organization.skills.' + stat),
+        title: game.i18n.format("KNW.Organization.skills.Test.Title", {
+          skill: game.i18n.localize("KNW.Organization.skills." + stat)
         }),
         content: `<label class="orgChooseActorLabel">${game.i18n.localize(
-          'KNW.Organization.skills.Test.ChooseActor'
+          "KNW.Organization.skills.Test.ChooseActor"
         )}
         <select class='orgChooseActor'>
         ${Handlebars.helpers.selectOptions(selectOptions, {
-          hash: { nameAttr: 'memberID', labelAttr: 'memberName' },
-        })}
+    hash: {nameAttr: "memberID", labelAttr: "memberName"}
+  })}
         </select></label>
         <label class="flexrow"><p class="orgUseProf">${game.i18n.localize(
-          'KNW.Organization.skills.Test.UseProf'
-        )}</p>
+    "KNW.Organization.skills.Test.UseProf"
+  )}</p>
         <input class="orgTestUseProf" type="checkbox" /></label>
         <p><em>${assocSkillsText}</em></p>`,
         buttons: {
           default: {
-            icon: '<i class="fa-solid fa-floppy-disk"></i>',
-            label: game.i18n.localize('KNW.Organization.skills.Test.Roll'),
+            icon: "<i class=\"fa-solid fa-floppy-disk\"></i>",
+            label: game.i18n.localize("KNW.Organization.skills.Test.Roll"),
             callback: (html) => {
               return {
                 chosenActor: game.actors.get(
-                  html.find('.orgChooseActor')[0].value
+                  html.find(".orgChooseActor")[0].value
                 ),
-                useProf: html.find('.orgTestUseProf')[0].checked,
+                useProf: html.find(".orgTestUseProf")[0].checked
               };
-            },
-          },
-        },
+            }
+          }
+        }
       });
       if (testInput?.chosenActor)
         this.actor.system.rollSkillTest(
@@ -220,40 +221,40 @@ export default class OrganizationSheet extends ActorSheet {
   }
 
   async #cyclePowerDie(event) {
-    const memberID = event.currentTarget.closest('li').dataset.id;
+    const memberID = event.currentTarget.closest("li").dataset.id;
     const currentValue = this.actor.system.powerPool[memberID];
     const memberActor = game.actors.get(memberID);
     switch (currentValue) {
       case null: // Available
         if (memberActor.isOwner) this.actor.system.rollPowerDie(memberID);
         else
-          ui.notifications.warn('KNW.Organization.Powers.Warning.MustOwnRoll', {
-            localize: true,
+          ui.notifications.warn("KNW.Organization.Powers.Warning.MustOwnRoll", {
+            localize: true
           });
         break;
       case 0: // Take Extended Rest
         if (memberActor.isOwner)
-          this.actor.update({ [`system.powerPool.${memberID}`]: null });
+          this.actor.update({[`system.powerPool.${memberID}`]: null});
         else
-          ui.notifications.warn('KNW.Organization.Powers.Warning.MustOwnRest', {
-            localize: true,
+          ui.notifications.warn("KNW.Organization.Powers.Warning.MustOwnRest", {
+            localize: true
           });
         break;
       default:
         ChatMessage.create({
           user: game.user,
-          content: game.i18n.format('KNW.Organization.Powers.TakeValue', {
+          content: game.i18n.format("KNW.Organization.Powers.TakeValue", {
             currentValue,
-            orgName: this.actor.name,
-          }),
+            orgName: this.actor.name
+          })
         });
-        this.actor.update({ [`system.powerPool.${memberID}`]: 0 });
+        this.actor.update({[`system.powerPool.${memberID}`]: 0});
     }
   }
 
   async #decrementPowerDie(event) {
     this.actor.system.decrementPowerDie(
-      event.currentTarget.closest('li').dataset.id
+      event.currentTarget.closest("li").dataset.id
     );
   }
 
@@ -261,28 +262,28 @@ export default class OrganizationSheet extends ActorSheet {
     return [
       {
         name: game.i18n.localize(
-          'KNW.Organization.Powers.ContextMenu.SetValue'
+          "KNW.Organization.Powers.ContextMenu.SetValue"
         ),
         icon: "<i class='fas fa-edit'></i>",
         condition: this.isEditable,
-        callback: this.setDieValue.bind(this),
+        callback: this.setDieValue.bind(this)
       },
       {
         name: game.i18n.localize(
-          'KNW.Organization.Powers.ContextMenu.ViewMember'
+          "KNW.Organization.Powers.ContextMenu.ViewMember"
         ),
         icon: "<i class='fas fa-eye'></i>",
         condition: true,
-        callback: this.viewMember.bind(this),
+        callback: this.viewMember.bind(this)
       },
       {
         name: game.i18n.localize(
-          'KNW.Organization.Powers.ContextMenu.RemoveMember'
+          "KNW.Organization.Powers.ContextMenu.RemoveMember"
         ),
         icon: "<i class='fas fa-trash'></i>",
         condition: this.isEditable,
-        callback: this.deleteMember.bind(this),
-      },
+        callback: this.deleteMember.bind(this)
+      }
     ];
   }
 
@@ -290,25 +291,25 @@ export default class OrganizationSheet extends ActorSheet {
     const memberID = html[0].dataset.id;
 
     new Dialog({
-      title: game.i18n.localize('KNW.Organization.Powers.SetValue'),
+      title: game.i18n.localize("KNW.Organization.Powers.SetValue"),
       content: Handlebars.helpers.numberInput(
         this.actor.system.powerPool[memberID],
-        { hash: { class: 'powerDie', min: 0, max: this.actor.system.powerDie } }
+        {hash: {class: "powerDie", min: 0, max: this.actor.system.powerDie}}
       ).string,
       buttons: {
         default: {
-          icon: '<i class="fa-solid fa-floppy-disk"></i>',
-          label: game.i18n.localize('Save Changes'),
+          icon: "<i class=\"fa-solid fa-floppy-disk\"></i>",
+          label: game.i18n.localize("Save Changes"),
           callback: (dialogHTML) => {
-            const newValue = dialogHTML.find('.powerDie')[0].value
-              ? dialogHTML.find('.powerDie')[0].value
+            const newValue = dialogHTML.find(".powerDie")[0].value
+              ? dialogHTML.find(".powerDie")[0].value
               : null;
             this.actor.update({
-              [`system.powerPool.${memberID}`]: newValue,
+              [`system.powerPool.${memberID}`]: newValue
             });
-          },
-        },
-      },
+          }
+        }
+      }
     }).render(true);
   }
 
@@ -322,11 +323,11 @@ export default class OrganizationSheet extends ActorSheet {
     const memberID = html[0].dataset.id;
     const member = game.actors.get(memberID);
     ui.notifications.info(
-      game.i18n.format('KNW.Organization.Powers.Warning.RemoveMember', {
+      game.i18n.format("KNW.Organization.Powers.Warning.RemoveMember", {
         memberName: member?.name,
-        organization: this.actor.name,
+        organization: this.actor.name
       })
     );
-    this.actor.update({ [`system.powerPool.-=${memberID}`]: null });
+    this.actor.update({[`system.powerPool.-=${memberID}`]: null});
   }
 }